requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
joblib==1.2.0
tensorflow-macos==2.13.0  # use tensorflow == 2.13.0 for windows or linux
opencv-python==4.7.0.68
